rutasycontroladores

+++ app/Http/Controllers/UserController.php
@@ -10,63 +10,68 @@ use Illuminate\Support\Facades\Hash;
 class UserController extends Controller
 {
     public function index(Request $req){
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
         return User::all();
     }
 
-    public function get($user){
-        $result=User::find($user);
-        //$result = DB::table('users')-> where('user','=',$user)->get();
+    public function get(Request $req, $user){
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
+        $result = User::find($user);
+        //$result = DB::table('users')->where('user', '=', $user)->get();
         if($result)
             return $result;
         else
-            return response()->json(['status'=>'failed'],404);
+            return response()->json(['status'=>'failed'], 404);
     }
 
     public function create(Request $req){
-        $this->validate($req,[
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
+        $this->validate($req, [
             'user'=>'required', 
             'nombre'=>'required',
             'pass'=>'required',
             'rol'=>'required']);
+
         $datos = new User;
-        //$datos->user=$req->user;
-        $datos->pass= Hash::make($req->pass);
-        //$datos->nombre=$req->nombre;
-        //$datos->rol=$req->rol;
-        //$datos->save();
+        // $datos->user = $req->user;
+        $datos->pass = Hash::make( $req->pass );
+        // $datos->nombre = $req->nombre;
+        // $datos->rol = $req->rol;
+        // $datos->save();
         $result = $datos->fill($req->all())->save();
-        if($result){
-            return response()->json(['status'=>'success'],200);
-        }else{
-            return response()->json(['status'=>'failed'],404);
-        }
+        if($result)
+            return response()->json(['status'=>'success'], 200);
+        else
+            return response()->json(['status'=>'failed'], 404);
     }
 
     public function update(Request $req, $user){
-        $this->validate($req,[
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
+        $this->validate($req, [
             'user'=>'filled', 
             'nombre'=>'filled',
             'pass'=>'filled',
             'rol'=>'filled']);
+
         $datos = User::find($user);
-        //$datos->pass=$req->pass;
+        $datos->pass = Hash::make( $req->pass );
+        if(!$datos) return response()->json(['status'=>'failed'], 404);
         $result = $datos->fill($req->all())->save();
-        if($result){
-            return response()->json(['status'=>'success'],200);
-        }else{
-            return response()->json(['status'=>'failed'],404);
-        }
+        if($result)
+            return response()->json(['status'=>'success'], 200);
+        else
+            return response()->json(['status'=>'failed'], 404);
     }
 
-    public function destroy($user){
-       
+    public function destroy(Request $req, $user){
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
         $datos = User::find($user);
-        if(!$datos) return response()->json(['status'=>'failed'],404);
+        if(!$datos) return response()->json(['status'=>'failed'], 404);
         $result = $datos->delete();
-        if($result){
-            return response()->json(['status'=>'success'],200);
-        }else{
-            return response()->json(['status'=>'failed'],404);
-        }
+        if($result)
+            return response()->json(['status'=>'success'], 200);
+        else
+            return response()->json(['status'=>'failed'], 404);
     }
+
 }
